- v1.75 - added failsafe entry to grub configuration (#334712) - added resume=/dev/<swap-part> to grub configuration (#335522) - remove unneeded /image directory after deployment (#335517) - fixed resetInstallationSource for zypper manager (#335134) - remove config-cdroot.tgz and config-cdroot.sh from image (#334645) - fixed interrupt signal handling for log process (#335588) - added "ZYPP_MODALIAS_SYSFS=/tmp" for zypper install/update (#335116) - use --auto-agree-with-licenses option for zypper in and outside   of the chroot environment (#335147)
git-svn-id: https://svn.berlios.de/svnroot/repos/kiwi/kiwi-head@723 92bc7f66-5d17-0410-ac0d-e081219a1bb3

+++ modules/KIWIImage.pm
@@ -866,6 +866,21 @@ sub createImageLiveCD {
 			return undef;
 		}
 		$kiwi -> done();
+		$kiwi -> info ("Removing CD root tarball from system image");
+		$data = qx (rm $cdrootData);
+		$code = $? >> 8;
+		if ($code != 0) {
+			$kiwi -> failed ();
+			$kiwi -> error  ($data);
+			$kiwi -> failed ();
+			if (! -d $main::RootTree.$baseSystem) {
+				qx (rm -rf $main::RootTree);
+				qx (rm -rf $tmpdir);
+				qx (rm -rf $imageTreeReadOnly);
+			}
+			return undef;
+		}
+		$kiwi -> done();
 	}
 	#==========================================
 	# Check for optional config-cdroot.sh
@@ -888,6 +903,21 @@ sub createImageLiveCD {
 			chomp $data;
 			$kiwi -> failed ();
 			$kiwi -> error  ("Failed to call CD root script: $data");
+			$kiwi -> failed ();
+			if (! -d $main::RootTree.$baseSystem) {
+				qx (rm -rf $main::RootTree);
+				qx (rm -rf $tmpdir);
+				qx (rm -rf $imageTreeReadOnly);
+			}
+			return undef;
+		}
+		$kiwi -> done();
+		$kiwi -> info ("Removing CD root setup script from system image");
+		$data = qx (rm $cdrootData);
+		$code = $? >> 8;
+		if ($code != 0) {
+			$kiwi -> failed ();
+			$kiwi -> error  ($data);
 			$kiwi -> failed ();
 			if (! -d $main::RootTree.$baseSystem) {
 				qx (rm -rf $main::RootTree);
